split transformerStore into separate files in a subdir
