+++ src/transformers/partition.ts
@@ -16,7 +16,7 @@ import {
 } from "../components/transformer-template/TransformerTemplate";
 import { readableName } from "../transformers/util";
 import { applyNewDataSet } from "../components/transformer-template/util";
-import { ActionTypes } from "../store/transformerStore";
+import { ActionTypes } from "../store/transformerStore/types";
 
 /**
  * Contains a dataset as a result of a partition, and the distinct
